+++ Editor/Constants.cs
@@ -1,0 +1,10 @@
+#nullable enable
+
+namespace asmdef2pu
+{
+    internal static class Constants
+    {
+        internal const string UnityEngineUiName = "UnityEngine.UI";
+        internal const string AssemblyCSharpName = "Assembly-CSharp";
+    }
+}